% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapper_sample.R
\name{dapper_chain}
\alias{dapper_chain}
\title{single chain sample}
\usage{
dapper_chain(
  data_model,
  sdp,
  init_par,
  niter = 2000,
  warmup = floor(niter/2),
  prg_bar = NULL
)
}
\description{
single chain sample
}
