% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_gaussian.R
\name{DiscreteGuassian}
\alias{DiscreteGuassian}
\alias{ddnorm}
\alias{rdnorm}
\title{The Discrete Gaussian Distribution}
\usage{
ddnorm(x, mu = 0, sigma = 1, log = FALSE)

rdnorm(n, mu = 0, sigma = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{mu}{location parameter.}

\item{sigma}{scale parameter.}

\item{log}{logical; if \code{TRUE}, probabilities are given as log(p).}

\item{n}{number of random deviates.}
}
\value{
ddnorm gives the probability mass function and rdnorm
generates random deviates.
}
\description{
The Discrete Gaussian Distribution
}
\details{
Probability mass function
\deqn{
P[X = x] = \dfrac{e^{-(x - \mu)^2/2\sigma^2}}{\sum_{y \in \mathbb{Z}} e^{-(x-\mu)^2/2\sigma^2}}.
}
}
\references{
Canonne, C. L., Kamath, G., & Steinke, T. (2020). The Discrete Gaussian for Differential Privacy.
\emph{arXiv}. \url{https://doi.org/10.48550/ARXIV.2004.00010}
}
