% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_laplace.R
\name{DiscreteLaplace}
\alias{DiscreteLaplace}
\alias{ddlaplace}
\alias{rdlaplace}
\title{Discrete Laplace Distribution}
\usage{
ddlaplace(x, scale = 1, log = FALSE)

rdlaplace(n, scale = 1)
}
\arguments{
\item{x}{a vector of quantiles.}

\item{scale}{the scale parameter.}

\item{log}{logical; if TRUE, probabilities are given as log(p).}

\item{n}{number of random deviates.}
}
\value{
dnorm gives the probability mass function and rdnorm
generates random deviates.
}
\description{
Discrete Laplace Distribution
}
