% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_sample.R
\name{gdp_chain}
\alias{gdp_chain}
\title{single chain sample}
\usage{
gdp_chain(
  data_model,
  sdp,
  nobs,
  init_par,
  niter = 2000,
  warmup = floor(niter/2),
  prg_bar = NULL
)
}
\description{
single chain sample
}
