% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_sample.R
\name{gdp_sample}
\alias{gdp_sample}
\title{Sample from posterior given private data}
\usage{
gdp_sample(
  data_model,
  sdp,
  nobs,
  init_par,
  niter = 2000,
  warmup = floor(niter/2),
  varnames = NULL
)
}
\arguments{
\item{data_model}{A data model created using the privacy class.}

\item{sdp}{The observed privatized data.}

\item{nobs}{The number of observations.}

\item{init_par}{Initial starting point of the chain.}

\item{niter}{Number of draws.}

\item{warmup}{Number of iterations to discard as warmup. Default is half of niter.}

\item{varnames}{Optional character vector specifying parameter names in the output.}

\item{chains}{Number of chains. The default is 1.}
}
\value{
A matrix of posterior samples.
}
\description{
Sample from posterior given private data
}
