% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_privacy.R
\name{mcmc_privacy}
\alias{mcmc_privacy}
\title{Sample from posterior}
\usage{
mcmc_privacy(data_model, sdp, nobs, init_par, niter)
}
\arguments{
\item{data_model}{A data model created using the privacy class.}

\item{sdp}{The observed privatized data.}

\item{nobs}{The number of observations.}

\item{init_par}{Initial starting point of the chain.}

\item{niter}{Number of draws.}
}
\value{
A matrix of posterior samples.
}
\description{
Sample from posterior
}
\examples{
...
}
