% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpsim_class.R
\name{new_dpout}
\alias{new_dpout}
\title{Creates simulation object}
\usage{
new_dpout(theta, accept_mat, varnames = NULL)
}
\arguments{
\item{theta}{}
}
\value{
a simulation object
}
\description{
Creates simulation object
}
