% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdpout_class.R
\name{new_gdpout}
\alias{new_gdpout}
\title{Creates simulation object}
\usage{
new_gdpout(theta, accept_mat, varnames = NULL)
}
\arguments{
\item{theta}{}
}
\value{
a simulation object
}
\description{
Creates simulation object
}
