% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapper_sample.R
\name{plot.dpout}
\alias{plot.dpout}
\title{Plot dpout object.}
\usage{
\method{plot}{dpout}(x, ...)
}
\arguments{
\item{x}{dp_out object.}

\item{...}{optional arguments to \code{mcmc_trace()}.}
}
\value{
trace plots.
}
\description{
Plot dpout object.
}
