% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapper_sample.R
\name{summary.dpout}
\alias{summary.dpout}
\title{Summarise dpout object.}
\usage{
\method{summary}{dpout}(object, ...)
}
\arguments{
\item{object}{dp_out object}

\item{...}{Optional arguments to summarise_draws.}
}
\value{
summary table
}
\description{
Summarise dpout object.
}
