% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_sample.R
\name{summary.gdpout}
\alias{summary.gdpout}
\title{Summarise dpout object.}
\usage{
\method{summary}{gdpout}(object)
}
\arguments{
\item{dpout}{object}
}
\value{
summary table
}
\description{
Summarise dpout object.
}
