% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_way.R
\name{two_way}
\alias{two_way}
\title{DP aware inference for two way table.}
\usage{
two_way(cell_values, n_total, rho, niter = 100, chains = 1)
}
\arguments{
\item{init_par}{}
}
\value{
p1,p2
}
\description{
DP aware inference for two way table.
}
\examples{
set.seed(1)
rho <- 3.325 * 3.01 * 10^(-2)
tval <- stats::rmultinom(1, 300, c(.20,.15,.35,.30))
sdp <- sapply(tval, function(s) rdpnorm(1, 0, sqrt(1/rho))) + tval
sout <- two_way(sdp, sum(tval), rho, 8000)
plot(sout)
}
