% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_laplace.R
\name{ddlaplace}
\alias{ddlaplace}
\title{Discrete Laplace Distribution}
\usage{
ddlaplace(x, scale = 1, log = FALSE)
}
\arguments{
\item{x}{a vector of quantiles.}

\item{scale}{the scale parameter.}

\item{log}{logical; if TRUE, probabilities are given as log(p).}
}
\value{
dnorm gives the probability mass function and rdnorm
generates random deviates.
}
\description{
Discrete Laplace Distribution
}
