% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_gaussian.R
\name{ddnorm}
\alias{ddnorm}
\title{The Discrete Gaussian Distribution}
\usage{
ddnorm(x, mu = 0, sigma = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{log}{logical; if TRUE, probabilities are given as log(p).}
}
\value{
dnorm gives the probability mass function and rdnorm
generates random deviates.
}
\description{
The Discrete Gaussian Distribution
}
