% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_laplace.R
\name{DiscreteLaplace}
\alias{DiscreteLaplace}
\alias{ddlaplace}
\alias{rdlaplace}
\title{Discrete Laplace Distribution}
\usage{
ddlaplace(x, scale = 1, log = FALSE)

rdlaplace(n, scale = 1)
}
\arguments{
\item{x}{a vector of quantiles.}

\item{scale}{the scale parameter.}

\item{log}{logical; if \code{TRUE}, probabilities are given as log(p).}

\item{n}{number of random deviates.}
}
\value{
ddlaplace gives the probability mass function and rdlaplace
generates random deviates.
}
\description{
Discrete Laplace Distribution
}
\details{
Probability mass function
\deqn{
P[X=x] = \dfrac{e^{1/t} - 1}{e^{1/t} + 1} e^{-|x|/t}.
}
}
\examples{
# mass function
ddlaplace(0)

# mass function is vectorized
ddlaplace(0:10, scale = 5)

# generate random samples
rdlaplace(10)

}
\references{
Canonne, C. L., Kamath, G., & Steinke, T. (2020). The Discrete Gaussian for Differential Privacy.
\emph{arXiv}. \url{https://doi.org/10.48550/ARXIV.2004.00010}
}
