% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpout.R
\name{new_dpout}
\alias{new_dpout}
\title{Creates dpout object.}
\usage{
new_dpout(theta, accept_mat, varnames = NULL)
}
\arguments{
\item{theta}{parameter}

\item{accept_mat}{acceptance matrix.}

\item{varnames}{variable names.}
}
\value{
a simulation object
}
\description{
Creates dpout object.
}
