% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_class.R
\name{new_privacy}
\alias{new_privacy}
\title{Creates data model}
\usage{
new_privacy(
  post_f = NULL,
  latent_f = NULL,
  priv_f = NULL,
  st_f = NULL,
  npar = NULL,
  varnames = NULL
)
}
\arguments{
\item{post_f}{a function that draws posterior samples given the confidential data.}

\item{latent_f}{a function that represents the latent data sampling model.}

\item{priv_f}{a function that represents the log likelihood of the privacy mechanism.}

\item{st_f}{a function that calculated the statistic to be released.}

\item{npar}{number of parameters in model.}

\item{varnames}{an optional character vector of parameter names. Used to label summary outputs.}
}
\value{
A data model of class privacy. Is a S3 object.
}
\description{
Creates data model
}
