% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_class.R
\name{new_privacy}
\alias{new_privacy}
\title{Creates data model}
\usage{
new_privacy(post_smpl, lik_smpl, ll_priv_mech, st_update, st_init, npar)
}
\arguments{
\item{post_smpl}{A function that draws posterior samples.}

\item{lik_smpl}{The sampling model.}

\item{ll_priv_mech}{The log likelihood of the privacy mechanism.}

\item{st_update}{Update function for statistic.}

\item{st_init}{Starting value of st.}

\item{npar}{Number of parameters.}
}
\value{
A data model of class privacy. Is a S3 object.
}
\description{
Creates data model
}
\examples{
...
}
