% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_class.R
\name{new_privacy}
\alias{new_privacy}
\title{Creates data model}
\usage{
new_privacy(
  post_smpl = NULL,
  lik_smpl = NULL,
  ll_priv_mech = NULL,
  st_calc = NULL,
  add = FALSE,
  npar = NULL,
  varnames = NULL
)
}
\arguments{
\item{post_smpl}{A function that draws posterior samples given the confidential data.}

\item{lik_smpl}{The sampling model.}

\item{ll_priv_mech}{The log likelihood of the privacy mechanism modulo an additive constant.}

\item{st_calc}{calculate statistic to be released.}

\item{npar}{Number of parameters in model.}
}
\value{
A data model of class privacy. Is a S3 object.
}
\description{
Creates data model
}
