% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapper_sample.R
\name{plot.dpout}
\alias{plot.dpout}
\title{Summarise dpout object.}
\usage{
\method{plot}{dpout}(x, ...)
}
\arguments{
\item{x}{dp_out object.}

\item{...}{Optional arguments to mcmc_trace.}
}
\value{
trace plots
}
\description{
Summarise dpout object.
}
