% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapper_sample.R
\name{summary.dpout}
\alias{summary.dpout}
\title{Summarise dpout object.}
\usage{
\method{summary}{dpout}(object, ...)
}
\arguments{
\item{object}{dp_out object}

\item{...}{optional arguments to \code{summarise_draws()}.}
}
\value{
A \code{\link[posterior:draws_summary]{draws_summary}} tibble containing
summary statistics (such as \code{mean}, \code{sd}, \code{ess_bulk}, and \code{rhat}).

See \code{\link[posterior:summarise_draws]{posterior::summarise_draws()}} for
details on how these statistics are calculated.
}
\description{
Summarise dpout object.
}
