% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapper_sample.R
\name{summary.dpout}
\alias{summary.dpout}
\title{Summarise dpout object.}
\usage{
\method{summary}{dpout}(object, ...)
}
\arguments{
\item{object}{dp_out object}

\item{...}{optional arguments to \code{summarise_draws()}.}
}
\value{
a summary table of MCMC statistics.
}
\description{
Summarise dpout object.
}
