% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_way.R
\name{two_way}
\alias{two_way}
\title{DP aware inference for two way table.}
\usage{
two_way(cell_values, n_total, epsilon, niter = 100)
}
\arguments{
\item{init_par}{}
}
\value{
p1,p2
}
\description{
DP aware inference for two way table.
}
\examples{
set.seed(1)
epsilon <- .02
tval <- stats::rmultinom(1, 300, c(.20,.15,.35,.30))
sdp <- sapply(tval, function(s) extraDistr::rlaplace(1,mu = s, sigma = 1/epsilon))
sout <- two_way(sdp, sum(tval), epsilon, 5000)
plot(sout)
}
