% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_way.R
\name{two_way}
\alias{two_way}
\title{DP aware inference for two way table.}
\usage{
two_way(cell_values, n_total, epsilon, niter = 100)
}
\arguments{
\item{init_par}{}
}
\value{
p1,p2
}
\description{
DP aware inference for two way table.
}
